requirements.txt
nicegui
keyring
pystray
Pillow
urllib3
beautifulsoup4
httpx
